src/modules/LlamaChat/requirements.macos.txt
#! Python3.7

huggingface_hub     # Installing the huggingface hub

diskcache>=5.6.1    # Installing disckcache for Disk and file backed persistent cache
numpy>=1.20.0       # Installing NumPy, a package for scientific computing

# "69008 Floating point exception: 8" on Intel macOS 11
# extra index could be: /basic, /AVX, /AVX2, or /AVX512
# --prefer-binary
# --extra-index-url https://jllllll.github.io/llama-cpp-python-cuBLAS-wheels/basic/cpu/

# Metal is not working on Intel macOS
# "unsupported op" on Intel macOS 11.0 on both metal and cpu
# --prefer-binary
# --extra-index-url https://abetlen.github.io/llama-cpp-python/whl/metal/

# --extra-index-url https://github.com/abetlen/llama-cpp-python/releases
--extra-index-url https://abetlen.github.io/llama-cpp-python/whl/cpu/
--prefer-binary
llama-cpp-python    # Installing simple Python bindings for the llama.cpp library

# last line empty